bot/requirements.txt
discord.py
python-dotenv
pymongo[srv]